Fix bug due to multi-lines and include new test script
There was a little bug due to multi-line entries, but this commit fixes
that. There's also a new script in the repo (parse_clan2_check.py)
that makes sure the output from the clan freq command corresponds to
the output of the parse_clan2 script. If there are entries missing in
the parse_clan2 output, the test script will indicate which ones
specifically (helpful when there's hundreds of words and only 1 or 2
missing). This will help track down issues with the parse_clan2 script
itself, as well as any formatting errors in the original .cha entries
which prevented the script from pulling out those missing words.

The test script takes 2 arguments:

path to clan freq output
path to parse_clan2 output

The clan freq output is just the list of words/counts copy and pasted
into a .txt file. There are examples in the data directory. I just ran
parse_clan2 and the test script over 02_12, 02_13, and 12_11. The only
missing words ended up being those which weren't properly formatted in
the .cha file. After changing them to have the correct format, the
script included them in the output csv.