Move error reporting to a final export file
Instead of having the script stop in the middle of parsing and fail
whenever there's a mistake in the entries, it will save all the
mistakes and output an error file at the end. This way, the user can
fix all the mistakes at once rather than running the script after each
mistake correction. The error output file is written to the same
directory as the input .cha file.